Plugin that implements a table to render labels in change screen
This plugin adds a user preference that allows users to choose a table
control to render the labels/approvals on the change screen (similar
to how labels/approvals were rendered on the old change screen).

Some users prefer the table view because it's easier to see which user
has already voted and whether there are users that haven't voted yet.

Change-Id: I57e8d5137f95f9a2553b2090ee0cb911fafc4c6b
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>