Add option to see reviewers with no votes under the default control
Adds an additional row to the default control for labels/approvals
that shows the reviewers that have not voted.

Some users like to know which reviewers haven't voted yet, but with
the default label control this is difficult to see, because it only
shows reviewers that have voted and the set of these users must be
compared with the set of all reviewers. Showing the set of reviewers
that has not voted yet as an extra row shows this information at a
glance.

Change-Id: I4cc24b374f7bb08f9d9ef824eace42bcb2b53d7b
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>